--- src/components/molecules/NewsArticleCard.js
@@ -1,35 +1,0 @@
-import { Card, Text, Image, Link } from '@chakra-ui/react'
-import pt from 'prop-types'
-
-export function NewsArticleCard({
-    title = 'We gotta save John Wic from Fortnite',
-    thumbnail = 'https://images.gog-statics.com/0aca7f0078df5a2d3d66a3122be6f93b015e98c1d85b0e5a0a8a75b94c748ce2.jpg',
-    link,
-}) {
-    return (
-        <Link href={link} isExternal>
-            <Card
-                data-testid="news-article-card"
-                bg="lightgrey"
-                overflow="hidden"
-            >
-                <Text p="5px" fontWeight="600">
-                    {title}
-                </Text>
-
-                <Image
-                    src={thumbnail}
-                    alt="missing image"
-                    w="250px"
-                    h="125px"
-                />
-            </Card>
-        </Link>
-    )
-}
-
-NewsArticleCard.propTypes = {
-    title: pt.string,
-    thumbnail: pt.string,
-    link: pt.string,
-}
