+++ src/components/molecules/NewsArticleCard/NewsArticleCard.module.scss
@@ -1,0 +1,14 @@
+.container {
+  background: lightgrey;
+
+  p{
+    padding: 5px;
+    font-weight: 600;
+  }
+
+  img{
+    width: 250px;
+    height: 125px;
+    object-fit: fill;
+  }
+}
