+++ src/components/molecules/ScrollSection/ScrollSection.module.scss
@@ -1,0 +1,10 @@
+.container{
+  display: flex;
+  flex-direction: column;
+  margin-bottom: 20px;
+
+  div{
+    display: flex;
+    justify-content: space-between;;
+  }
+}
