+++ src/components/molecules/index.js
@@ -1,3 +1,3 @@
-export { GameCard } from './GameCard'
-export { ScrollSection } from './ScrollSection'
-export { NewsArticleCard } from './NewsArticleCard'
+export { GameCard } from './GameCard/GameCard'
+export { ScrollSection } from './ScrollSection/ScrollSection'
+export { NewsArticleCard } from './NewsArticleCard/NewsArticleCard'
